+++ source/LeLib/limage/limagebbc.cpp
@@ -112,7 +112,7 @@ unsigned int LImageBBC::getPixel(int x, int y)
     return 0;
 }
 */
-void LImageBBC::ToQImage(LColorList &lst, QImage &img, float zoom, QPointF center)
+void LImageBBC::ToQImage(LColorList &lst, QImage &img, double zoom, QPointF center)
 {
 #pragma omp parallel for
     for (int i=0;i<m_width;i++)
